Extract default replaceInvalidValue from GroupingSelector
